# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Autoken Facturas v2",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/uploads": {
      "post": {
        "tags": [
          "intake"
        ],
        "summary": "Upload File",
        "description": "Sube un fichero de factura a una empresa. Ver spec S2.1 para los c\u00f3digos (201/4xx/503).\n\nOrden: pertenencia (403/404) -> tama\u00f1o (413) -> el servicio hace el resto (415/422/409/503/201).\n\n`sharpness_score` (S6.14 C8): nitidez de la captura calculada en cliente (varianza del\nLaplaciano), string opcional. Es TELEMETR\u00cdA, no dato de dominio: no se persiste ni se valida\n(un valor raro no rompe la subida); solo se loguea como m\u00e9trica de calidad de captura.",
        "operationId": "upload_file_api_v1_uploads_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_v1_uploads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/uploads/{file_id}/draft": {
      "put": {
        "tags": [
          "invoicing"
        ],
        "summary": "Save Review Draft",
        "description": "Guarda un borrador con control optimista de revisi\u00f3n (R-022).",
        "operationId": "save_review_draft_api_v1_uploads__file_id__draft_put",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDraftIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{file_id}/review": {
      "get": {
        "tags": [
          "invoicing"
        ],
        "summary": "Review Upload",
        "description": "Datos de revisi\u00f3n de un fichero ya le\u00eddo (S2.4). 200 con datos; 403/404/409 por acceso.",
        "operationId": "review_upload_api_v1_uploads__file_id__review_get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Review Upload Api V1 Uploads  File Id  Review Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{file_id}/counterparty-verdict": {
      "post": {
        "tags": [
          "invoicing"
        ],
        "summary": "Draft Counterparty Verdict",
        "description": "Valida el CIF/nombre que se est\u00e1 editando, sin guardar una factura (S6.10 C6).",
        "operationId": "draft_counterparty_verdict_api_v1_uploads__file_id__counterparty_verdict_post",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftCounterpartyVerdictIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftCounterpartyVerdictOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{file_id}/confirm": {
      "post": {
        "tags": [
          "invoicing"
        ],
        "summary": "Confirm Upload",
        "description": "Confirma un fichero y persiste su factura (S2.5). 201 con el id; 4xx seg\u00fan las guardas.",
        "operationId": "confirm_upload_api_v1_uploads__file_id__confirm_post",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Confirm Upload Api V1 Uploads  File Id  Confirm Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{file_id}/review-readonly": {
      "get": {
        "tags": [
          "invoicing"
        ],
        "summary": "Review Upload Readonly",
        "description": "Abre un pendiente ajeno para supervisi\u00f3n sin exponer acciones de escritura (R-026).",
        "operationId": "review_upload_readonly_api_v1_uploads__file_id__review_readonly_get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Review Upload Readonly Api V1 Uploads  File Id  Review Readonly Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoices/pending-supervision": {
      "get": {
        "tags": [
          "invoicing"
        ],
        "summary": "Pending Supervision",
        "description": "Pendientes de otros usuarios del tenant, solo metadata y paginaci\u00f3n estable (R-026).",
        "operationId": "pending_supervision_api_v1_invoices_pending_supervision_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupervisionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/invoices/inbox": {
      "get": {
        "tags": [
          "invoicing"
        ],
        "summary": "Invoice Inbox",
        "description": "Bandeja SELF ONLY, tambi\u00e9n para `tenant_admin`, con resumen y cursor estable.",
        "operationId": "invoice_inbox_api_v1_invoices_inbox_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboxOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_file_api_v1_uploads_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "recibida",
                  "emitida"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "sharpness_score": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharpness Score"
          }
        },
        "type": "object",
        "required": [
          "file",
          "company_id"
        ],
        "title": "Body_upload_file_api_v1_uploads_post"
      },
...
      "DraftCounterpartyVerdictIn": {
        "properties": {
          "counterparty_tax_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counterparty Tax Id"
          },
          "counterparty_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counterparty Name"
          }
        },
        "type": "object",
        "title": "DraftCounterpartyVerdictIn",
        "description": "Valores actuales del formulario para validar antes de confirmar (S6.10)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewDraftOut": {
        "properties": {
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "revision",
          "updated_at"
        ],
        "title": "ReviewDraftOut"
      },
...
    }
  }
}